+++ src/main/java/chess/service/PieceConvertor.java
@@ -1,6 +1,13 @@
-package chess.domain.piece;
+package chess.service;
 
 import chess.domain.Color;
+import chess.domain.piece.Bishop;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
 import java.util.Arrays;
 import java.util.List;
 import java.util.NoSuchElementException;
